Fix the cogen value system lookup.
