+++ src/codegen/type/integer_type.cpp
@@ -11,10 +11,8 @@
 //===----------------------------------------------------------------------===//
 
 #include "codegen/type/integer_type.h"
-#include "codegen/type/timestamp_type.h"
 
 #include "codegen/lang/if.h"
-#include "codegen/proxy/timestamp_functions_proxy.h"
 #include "codegen/proxy/values_runtime_proxy.h"
 #include "codegen/type/boolean_type.h"
 #include "codegen/value.h"
@@ -376,33 +374,6 @@ struct Modulo : public TypeSystem::BinaryOperator {
   }
 };
 
-// DateTrunc
-// TODO(lma): Move this to the Timestamp type once the function lookup logic is
-// corrected.
-struct DateTrunc : public TypeSystem::BinaryOperator {
-  bool SupportsTypes(const Type &left_type,
-                     const Type &right_type) const override {
-    return left_type.GetSqlType() == Integer::Instance() &&
-           right_type.GetSqlType() == Timestamp::Instance();
-  }
-
-  Type ResultType(UNUSED_ATTRIBUTE const Type &left_type,
-                  UNUSED_ATTRIBUTE const Type &right_type) const override {
-    return Type{Timestamp::Instance()};
-  }
-
-  Value DoWork(CodeGen &codegen, const Value &left, const Value &right,
-               UNUSED_ATTRIBUTE OnError on_error) const override {
-    PL_ASSERT(SupportsTypes(left.GetType(), right.GetType()));
-
-    // TODO(lma): I actually don't know whether this should be called
-    // TimestampFunctionsProxy or IntegerFunctionsProxy...
-    llvm::Value *raw_ret = codegen.Call(TimestampFunctionsProxy::DateTrunc,
-                                        {left.GetValue(), right.GetValue()});
-    return Value{Timestamp::Instance(), raw_ret};
-  }
-};
-
 //===----------------------------------------------------------------------===//
 // TYPE SYSTEM CONSTRUCTION
 //===----------------------------------------------------------------------===//
@@ -444,11 +415,13 @@ static Sub kSubOp;
 static Mul kMulOp;
 static Div kDivOp;
 static Modulo kModuloOp;
-static DateTrunc kDateTrunc;
 static std::vector<TypeSystem::BinaryOpInfo> kBinaryOperatorTable = {
-    {OperatorId::Add, kAddOp},    {OperatorId::Sub, kSubOp},
-    {OperatorId::Mul, kMulOp},    {OperatorId::Div, kDivOp},
-    {OperatorId::Mod, kModuloOp}, {OperatorId::DateTrunc, kDateTrunc}};
+    {OperatorId::Add, kAddOp},
+    {OperatorId::Sub, kSubOp},
+    {OperatorId::Mul, kMulOp},
+    {OperatorId::Div, kDivOp},
+    {OperatorId::Mod, kModuloOp},
+};
 
 // Nary operations
 static std::vector<TypeSystem::NaryOpInfo> kNaryOperatorTable = {};
