+++ src/codegen/type/varchar_type.cpp
@@ -17,6 +17,8 @@
 #include "codegen/proxy/values_runtime_proxy.h"
 #include "codegen/type/boolean_type.h"
 #include "codegen/type/integer_type.h"
+#include "codegen/type/timestamp_type.h"
+#include "codegen/proxy/timestamp_functions_proxy.h"
 
 namespace peloton {
 namespace codegen {
@@ -196,6 +198,31 @@ struct Length : public TypeSystem::UnaryOperator {
   }
 };
 
+// DateTrunc
+// TODO(lma): Move this to the Timestamp type once the function lookup logic is
+// corrected.
+struct DateTrunc : public TypeSystem::BinaryOperator {
+  bool SupportsTypes(const Type &left_type,
+                     const Type &right_type) const override {
+    return left_type.GetSqlType() == Varchar::Instance() &&
+           right_type.GetSqlType() == Timestamp::Instance();
+  }
+
+  Type ResultType(UNUSED_ATTRIBUTE const Type &left_type,
+                  UNUSED_ATTRIBUTE const Type &right_type) const override {
+    return Type{Timestamp::Instance()};
+  }
+
+  Value DoWork(CodeGen &codegen, const Value &left, const Value &right,
+               UNUSED_ATTRIBUTE OnError on_error) const override {
+    PL_ASSERT(SupportsTypes(left.GetType(), right.GetType()));
+
+    llvm::Value *raw_ret = codegen.Call(TimestampFunctionsProxy::DateTrunc,
+                                        {left.GetValue(), right.GetValue()});
+    return Value{Timestamp::Instance(), raw_ret};
+  }
+};
+
 //===----------------------------------------------------------------------===//
 // TYPE SYSTEM CONSTRUCTION
 //===----------------------------------------------------------------------===//
@@ -220,8 +247,9 @@ static std::vector<TypeSystem::UnaryOpInfo> kUnaryOperatorTable = {
 
 // Binary operations
 static Like kLike;
+static DateTrunc kDateTrunc;
 static std::vector<TypeSystem::BinaryOpInfo> kBinaryOperatorTable = {
-    {OperatorId::Like, kLike}};
+    {OperatorId::Like, kLike}, {OperatorId::DateTrunc, kDateTrunc}};
 
 // Nary operations
 static std::vector<TypeSystem::NaryOpInfo> kNaryOperatorTable = {};
